requirements.txt
django
whitenoise
openai
pillow
matplotlib
python-dotenv
twilio